+++ projects/swimlane/ngx-ui/src/lib/components/dropzone/dropzone.component.ts
@@ -10,6 +10,7 @@ import {
 import { coerceBooleanProperty } from '@angular/cdk/coercion';
 import { FileUploaderOptions, FileUploader, FileItem } from '@swimlane/ng2-file-upload';
 import { id } from '../../utils/id/id.util';
+import { listFormatter } from '../../utils/list-format/list-format.util';
 
 export enum DropzoneSize {
   Small = 'small',
@@ -50,8 +51,7 @@ export class DropzoneComponent implements OnInit {
 
   ngOnInit(): void {
     if (this.acceptedFileFormats && this.acceptedFileFormats.length) {
-      this.acceptedFileFormatsTextDisplay =
-        this.acceptedFileFormats.slice(0, -1).join(', ') + ' and ' + this.acceptedFileFormats.slice(-1);
+      this.acceptedFileFormatsTextDisplay = listFormatter.format(this.acceptedFileFormats);
     }
   }
 }