+++ projects/swimlane/ngx-ui/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## HEAD (unreleased)
 
+- Fix: Display proper supports message with only one allowed file format on `Dropzone` component
 - Fix: Fix various alignment issues in selects
 - Fix: select box with tagging enabled triggers form submit on enter